Move polygon_supernet_* files into a separate folder
